old_bot/requirements.txt
discord.py>=1.2.3
emoji
fuzzywuzzy
googletrans>=2.4.0
Pillow>=6.1.0
python-dotenv>=0.10.3
python-Levenshtein
requests>=2.22.0
Wikipedia-API
wolframalpha
xkcd
beautifulsoup4
latex
pdf2image
pypng

wheel
cairosvg
